pyproject.toml
[project]
name = "smart-calendar"
version = "1.0.0"
description = "Smart Email-to-Calendar Pipeline — watches Gmail, parses events with a free LLM, creates Google Calendar events with intelligent notifications"
requires-python = ">=3.10"
dependencies = [
    # Google APIs
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    # HTTP client (async)
    "httpx>=0.27.0",
    # Config / env loading
    "python-dotenv>=1.0.0",
    # Web server (for Vercel deployment + iOS Shortcuts)
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    # Pydantic (bundled with FastAPI, pinned for compatibility)
    "pydantic>=2.0.0",
    "whisper-cpp-cli>=0.0.3",
    "python-multipart>=0.0.22",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "transformers>=4.51.0",
    "huggingface-hub>=0.31.0",
    "torch>=2.7.0",
    "safetensors>=0.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",   # for TestClient
]

[project.scripts]
smart-calendar = "backend.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
